+++ src/Encina.Dapper.PostgreSQL/Tenancy/TenantAwareFunctionalRepositoryDapper.cs
@@ -526,7 +526,7 @@ public sealed class TenantAwareFunctionalRepositoryDapper<TEntity, TId> : IFunct
             if (entity is IVersionedEntity versionedEntity)
             {
                 var originalVersion = versionedEntity.Version;
-                versionedEntity.Version = (int)(originalVersion + 1);
+                versionedEntity.Version = originalVersion + 1;
 
                 var versionedSql = BuildVersionedUpdateSql();
 
@@ -737,7 +737,7 @@ public sealed class TenantAwareFunctionalRepositoryDapper<TEntity, TId> : IFunct
                     if (entity is IVersionedEntity versionedEntity)
                     {
                         var originalVersion = versionedEntity.Version;
-                        versionedEntity.Version = (int)(originalVersion + 1);
+                        versionedEntity.Version = originalVersion + 1;
 
                         var parameters = new DynamicParameters(entity);
                         parameters.Add("OriginalVersion", originalVersion);
