+++ src/Encina.MongoDB/Repository/FunctionalRepositoryMongoDB.cs
@@ -604,7 +604,7 @@ public sealed class FunctionalRepositoryMongoDB<TEntity, TId> : IFunctionalRepos
                 if (entity is IVersionedEntity versionedEntity)
                 {
                     originalVersions[entity] = versionedEntity.Version;
-                    versionedEntity.Version = (int)(versionedEntity.Version + 1);
+                    versionedEntity.Version = versionedEntity.Version + 1;
                 }
                 else if (entity is IVersioned versioned)
                 {
